refact(wasm): use helper module instead of banner
